refactor: combinnate watcher
