+++ packages/utils/src/index.ts
@@ -6,8 +6,7 @@ export * from './tools/common'
 export * from './tools/tool'
 export * from './polyfill/recover-native'
 export * from './tools/dom'
-export * from './scroll'
-export * from './output'
-export * from './transform'
+export * from './tools/scroll'
+export * from './tools/output'
+export * from './tools/transform'
 export * from './performance/fmp'
-export * from './tools/watcher'
